fix errros

+++ cron/stagin_full_after_part_stock_update_legacy.php
@@ -49,40 +49,43 @@ $contador  = 0;
 
 
 $sql = sprintf(
-    "SELECT `Staging Key` FROM `Staging Dimension`  where `Staging Operation`='full_after_part_stock_update_legacy' "
+    "SELECT `Staging Key`,`Staging Object Key` FROM `Staging Dimension`  where `Staging Operation`='full_after_part_stock_update_legacy' "
 );
 
 if ($result = $db->query($sql)) {
     foreach ($result as $row) {
         $part =  get_object('Part',$row['Staging Object Key']);
 
-
-        $editor['Date'] = gmdate('Y-m-d H:i:s');
-        $part->editor = $editor;
-
-        $part->activate();
-        $part->discontinue_trigger();
-
-
-        $part->update_available_forecast();
-        $part->update_stock_status();
-
-        foreach ($part->get_products('objects') as $product) {
+        if($part->id){
 
             $editor['Date'] = gmdate('Y-m-d H:i:s');
-            $product->editor = $editor;
+            $part->editor = $editor;
 
-            $product->fork = true;
-            $product->update_availability(false);
+            $part->activate();
+            $part->discontinue_trigger();
+
+
+            $part->update_available_forecast();
+            $part->update_stock_status();
+
+            foreach ($part->get_products('objects') as $product) {
+
+                $editor['Date'] = gmdate('Y-m-d H:i:s');
+                $product->editor = $editor;
+
+                $product->fork = true;
+                $product->update_availability(false);
+            }
+
+
+
+
+            $sql=sprintf('delete from `Staging Dimension`  where `Staging Key`=%d ',$row['Staging Key']);
+            $db->exec($sql);
         }
 
         $contador++;
         $lap_time1 = date('U');
-
-
-        $sql=sprintf('delete from `Staging Dimension`  where `Staging Key`=%d ',$row['Staging Key']);
-        $db->exec($sql);
-
 
         if ($print_est) {
             print 'Pa '.percentage($contador, $total, 3)."  lap time ".sprintf("%.2f", ($lap_time1 - $lap_time0) / $contador)." EST  ".sprintf(
